+++ webhook/main.py
@@ -1,9 +1,8 @@
-from enum import Enum
-from fastapi import FastAPI, Request, Header, HTTPException, Body
+from fastapi import FastAPI, Header, HTTPException
 from google.cloud import ndb
-from pydantic import BaseModel, validator
-from typing import List, Optional, Dict, Union
-from validate_email import validate_email
+from typing import Optional, Dict
+
+from models import VTAPI, EmailWrapper
 
 app = FastAPI()
 client = ndb.Client()
@@ -14,26 +13,6 @@ async def root():
     return {"data": "Hello World"}
 
 
-class VTType(str, Enum):
-    file = 'file'
-    url = 'url'
-    domain = 'domain'
-    ip_address = 'ip_address'
-
-
-class VTData(BaseModel):
-    attributes: Dict
-    id: str
-    links: Dict
-    type: VTType
-
-
-class VTAPI(BaseModel):
-    data: List[VTData]
-    links: Dict
-    meta: Dict
-
-
 @app.post("/query-results/")
 async def query_results(data: VTAPI, x_appengine_inbound_appid: Optional[str] = Header(None)):
     if x_appengine_inbound_appid != 'virustotal-step-2020':
@@ -42,29 +21,14 @@ async def query_results(data: VTAPI, x_appengine_inbound_appid: Optional[str] = 
     return data
 
 
-class Email(BaseModel):
-    email: str
-
-    @validator('email')
-    def email_validator(cls, v):
-        if not validate_email(email_address=v,
-                              check_regex=True,
-                              check_mx=True,
-                              debug=True):
-            raise ValueError("Email address is not valid!")
-        return v
-
-
 class UserEmail(ndb.Model):
     api_key: ndb.StringProperty
     email: ndb.StringProperty
 
 
-@app.put("/email-address/{api_key}")
-async def set_email(api_key: str, email: Email):
-    if not api_key.isalnum():
-        raise HTTPException(status_code=400, detail="API key must be alphanumeric!")
+@app.post("/email-address/")
+async def set_email(content: EmailWrapper):
 
-    return email
+    return content
     # with client.context():
     #     UserEmail(api_key=api_key, email=email).put()
